Further revisions inc. moving to trap to close port-forwarding

+++ logging/bin/user.sh
@@ -9,6 +9,9 @@ cd "$(dirname $BASH_SOURCE)/../.."
 source logging/bin/common.sh
 this_script=`basename "$0"`
 
+source logging/bin/rbac-include.sh
+source logging/bin/apiaccess-include.sh
+
 function show_usage {
   which_action=$1
   case "$which_action" in
@@ -16,9 +19,10 @@ function show_usage {
        log_info  ""
        log_info  "Usage: $this_script CREATE [REQUIRED_PARAMETERS] [OPTIONS] "
        log_info  ""
-       log_info  "Creates a user in the internal user database and grants the user permission to access log messages for the specified namespace"
+       log_info  "Creates a user in the internal user database and limits their access to only log messages from the specified namespace and, optionally, more narrowly, to the specified tenant within the specified namespace"
        log_info  ""
        log_info  "     -ns, --namespace   NAMESPACE - (Required) The Viya deployment/Kubernetes Namespace to which this user should be granted access"
+       log_info  "     -t,  --tenant      TENANT    - (Optional) The tenant within the specific Viya deployment/Kubernetes Namespace to which this user should be granted access."
        log_info  "     -u,  --user        USERNAME  - (Optional) The username to be created (defaults to match NAMESPACE)"
        log_info  "     -p,  --password    PASSWORD  - (Optional) The password for the newly created account (defaults to match USERNAME)"
        echo ""
@@ -34,7 +38,7 @@ function show_usage {
        log_info  ""
        log_info  "Usage: $this_script ACTION [REQUIRED_PARAMETERS] [OPTIONS] "
        log_info  ""
-       log_info  "Creates or deletes a user in the internal user database.  Newly created users are granted permission to access log messages for the specified namespace"
+       log_info  "Creates or deletes a user in the internal user database.  Newly created users are granted permission limiting their access to log messages for the specified namespace, and, optionally, to a specific tenant within that namespace."
        log_info  ""
        log_info  "        ACTION - (Required) one of the following actions: [CREATE, DELETE]"
        log_info  ""
@@ -61,6 +65,16 @@ while (( "$#" )); do
         log_error "Error: A value for parameter [Namespace] has not been provided." >&2
         show_usage
         exit 2
+      fi
+      ;;
+    -t|--tenant)
+      if [ -n "$2" ] && [ ${2:0:1} != "-" ]; then
+        TENANT=$2
+        shift 2
+      else
+        log_error "Error: A value for parameter [Tenant] has not been provided." >&2
+        show_usage
+      exit 2
       fi
       ;;
     -u|--username)
@@ -122,16 +136,21 @@ fi
 
 log_debug "POS_PARMS: $POS_PARMS"
 
+if [ -n "$TENANT" ]; then
+   NST="${NAMESPACE}_${TENANT}"
+else
+   NST="$NAMESPACE"
+fi
+
 if [ -z "$USERNAME"  ] && [ -z "$NAMESPACE" ]; then
   log_error "Required parameter(s) NAMESPACE and/or USERNAME not specified"
   show_usage $action
   exit 4
 elif [ -z "$USERNAME" ]; then
-  USERNAME="$NAMESPACE"
+  USERNAME="$NST"
 fi
 
-log_debug "NAMESPACE: $NAMESPACE USERNAME: $USERNAME"
-
+log_debug "NAMESPACE: $NAMESPACE TENANT: $TENANT NST: $NST USERNAME: $USERNAME"
 
 # get admin credentials
 export ES_ADMIN_USER=$(kubectl -n $LOG_NS get secret internal-user-admin -o=jsonpath="{.data.username}" |base64 --decode)
@@ -140,37 +159,18 @@ export ES_ADMIN_PASSWD=$(kubectl -n $LOG_NS get secret internal-user-admin -o=js
 #temp file to hold responses
 tmpfile=$TMP_DIR/output.txt
 
-# set up temporary port forwarding to allow curl access
-ES_PORT=$(kubectl -n $LOG_NS get service v4m-es-client-service -o=jsonpath='{.spec.ports[?(@.name=="http")].port}')
+# Get Security API URL
+get_sec_api_url
 
-# command is sent to run in background
-kubectl -n $LOG_NS port-forward --address localhost svc/v4m-es-client-service :$ES_PORT > $tmpfile  &
-
-# get PID to allow us to kill process later
-pfPID=$!
-log_debug "pfPID: $pfPID"
-
-# pause to allow port-forwarding messages to appear
-sleep 5s
-
-# determine which port port-forwarding is using
-pfRegex='Forwarding from .+:([0-9]+)'
-myline=$(head -n1  $tmpfile)
-
-if [[ $myline =~ $pfRegex ]]; then
-   TEMP_PORT="${BASH_REMATCH[1]}";
-   log_debug "TEMP_PORT=${TEMP_PORT}"
-else
-   set +e
-   log_error "Unable to obtain or identify the temporary port used for port-forwarding; exiting script.";
-   kill -9 $pfPID
-   rm -f  $tmpfile
-   exit 18
+if [ -z "$sec_api_url" ]; then
+   log_error "Unable to determine URL to access security API endpoint"
+   exit 1
 fi
 
 
+
 # Check if user exists
-response=$(curl -s -o /dev/null -w "%{http_code}" -XGET "https://localhost:$TEMP_PORT/_opendistro/_security/api/internalusers/$USERNAME"  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
+response=$(curl -s -o /dev/null -w "%{http_code}" -XGET "$sec_api_url/internalusers/$USERNAME"  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
 if [[ $response == 404 ]]; then
    USER_EXISTS=false
 else
@@ -184,52 +184,59 @@ case "$action" in
          log_error "Required argument NAMESPACE no specified"
          echo ""
          show_usage CREATE
-         kill -9 $pfPID
-         exit 2
+         exit 1
       fi
-      log_info "Attempting to create user [$USERNAME] and grant them access to namespace [$NAMESPACE] [$(date)]"
+      if [ -n "$TENANT" ]; then
+         NST="${NAMESPACE}_${TENANT}"
+         log_info "Attempting to create user [$USERNAME] and grant them access to the tenant [$TENANT] within the namespace [$NAMESPACE] [$(date)]"
+      else
+         NST="$NAMESPACE"
+         log_info "Attempting to create user [$USERNAME] and grant them access to namespace [$NAMESPACE] [$(date)]"
+      fi
+
+      # Check if user exists
+      if [[ "$USER_EXISTS" == "true" ]]; then
+         log_error "A user with this name [$USERNAME] already exists. [$response]"
+          exit 1
+      fi
 
       INDEX_PREFIX=viya_logs
-      ROLENAME=search_index_$NAMESPACE
+      ROLENAME=search_index_$NST
 
       #Check if role exists
-      response=$(curl -s -o /dev/null -w "%{http_code}" -XGET "https://localhost:$TEMP_PORT/_opendistro/_security/api/roles/$ROLENAME"  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
-
-      if [[ $response != 2* ]]; then
+      if ! role_exists $ROLENAME; then
          log_error "The expected access control role [$ROLENAME] does NOT exist; the role must be created before users can be linked to that role."
-         kill -9 $pfPID
-         exit 5
+         exit 1
       fi
 
       PASSWORD=${PASSWORD:-$USERNAME}
+      if [ -z "$TENANT" ]; then
+         TCONSTRAINT="-none-"
+      else
+         TCONSTRAINT=$TENANT
+      fi
 
       cp logging/es/odfe/rbac $TMP_DIR -r
 
       # Replace PLACEHOLDERS
       sed -i'.bak' "s/xxIDXPREFIXxx/$INDEX_PREFIX/g"  $TMP_DIR/rbac/*.json                  # IDXPREFIX
+      sed -i'.bak' "s/xxNSTxx/$NST/g"                 $TMP_DIR/rbac/*.json                  # NAMESPACE|NAMESPACE_TENANT
       sed -i'.bak' "s/xxNAMESPACExx/$NAMESPACE/g"     $TMP_DIR/rbac/*.json                  # NAMESPACE
+      sed -i'.bak' "s/xxTENANTxx/$TENANT/g"           $TMP_DIR/rbac/*.json                  # TENANT
+      sed -i'.bak' "s/xxTCONSTRAINTxx/$TCONSTRAINT/g" $TMP_DIR/rbac/*.json                  # TENANT|'-none-'
       sed -i'.bak' "s/xxPASSWORDxx/$PASSWORD/g"       $TMP_DIR/rbac/*.json                  # PASSWORD
       sed -i'.bak' "s/xxCREATEDBYxx/$this_script/g"   $TMP_DIR/rbac/*.json                  # CREATEDBY
       sed -i'.bak' "s/xxDATETIMExx/$(date)/g"         $TMP_DIR/rbac/*.json                  # DATE
 
       log_debug "Contents of user.json template file after substitutions: \n $(cat $TMP_DIR/rbac/user.json)"
 
-      # Check if user exists
-      if [[ "$USER_EXISTS" == "true" ]]; then
-         log_error "There was an issue creating the user [$USERNAME]; user already exists. [$response]"
-         kill -9 $pfPID
-         exit 19
-      else
-         log_debug "User [$USERNAME] does not exist. [$response]"
-      fi
 
       # Create user
-      response=$(curl -s -o /dev/null -w "%{http_code}" -XPUT "https://localhost:$TEMP_PORT/_opendistro/_security/api/internalusers/$USERNAME"  -H 'Content-Type: application/json' -d @$TMP_DIR/rbac/user${READONLY_FLAG}.json  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
+      response=$(curl -s -o /dev/null -w "%{http_code}" -XPUT "$sec_api_url/internalusers/$USERNAME"  -H 'Content-Type: application/json' -d @$TMP_DIR/rbac/user${READONLY_FLAG}.json  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
 
       if [[ $response != 2* ]]; then
          log_error "There was an issue creating the user [$USERNAME]. [$response]"
-         kill -9 $pfPID
-         exit 17
+         exit 1
       else
       log_info "User [$USERNAME] created. [$response]"
       fi
@@ -240,8 +247,7 @@ case "$action" in
          log_error "Required argument USERNAME not specified"
          echo ""
          show_usage DELETE
-         kill -9 $pfPID
-         exit 2
+         exit 1
       fi
 
       log_info "Attempting to remove user [$USERNAME] from the internal user database [$(date)]"
@@ -249,18 +255,16 @@ case "$action" in
       # Check if user exists
       if [[ "$USER_EXISTS" != "true" ]]; then
          log_error "There was an issue deleting the user [$USERNAME]; the user does NOT exists. [$response]"
-         kill -9 $pfPID
-         exit 20
+         exit 1
       else
          log_debug "User [$USERNAME] exists. [$response]"
       fi
 
       # Delete user
-      response=$(curl -s -o /dev/null -w "%{http_code}" -XDELETE "https://localhost:$TEMP_PORT/_opendistro/_security/api/internalusers/$USERNAME"  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
+      response=$(curl -s -o /dev/null -w "%{http_code}" -XDELETE "$sec_api_url/internalusers/$USERNAME"  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
       if [[ $response != 2* ]]; then
          log_error "There was an issue deleting the user [$USERNAME]. [$response]"
-         kill -9 $pfPID
-         exit 17
+         exit 1
       else
          log_info "User [$USERNAME] deleted. [$response]"
          log_notice "User [$USERNAME] removed from internal user database [$(date)]"
@@ -268,13 +272,11 @@ case "$action" in
       ;;
    *)
       log_error "Invalid action specified"
-      kill -9 $pfPID
-      exit 3
+      exit 1
    ;;
 esac
 
 
+#remove tmpfile
+rm -f $tmpfile
 
-# terminate port-forwarding and remove tmpfile
-log_info "You may see a message below about a process being killed; it is expected and can be ignored."
-kill  -9 $pfPID
